pyproject.toml
[project]
name = "search-subbrands"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pika>=1.3.2",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.0",
]
